removing base tests
